requirements.txt
discord.py==2.3.2
openai==1.3.8
rich==13.7.0